refactor: move helper code into utils
